Move unit tests to Factory class tests
